V2.0.0 stable after Hotfix

+++ functions/importLegacyGunDatabase.ts
@@ -110,7 +110,12 @@ export async function importLegacyGunDatabase(resizeImages:boolean, importOption
         }
     }))
 
-    await db.delete(schema.gunCollection);
+    if(importOptionLegacyDB === "gun"){
+        await db.delete(schema.gunCollection);
+    }
+    if(importOptionLegacyDB === "ammo"){
+        await db.delete(schema.ammoCollection);
+    }
 
    for(const importable of importableCollection){
     if(importOptionLegacyDB === "gun"){
